Recursively find deps
Unlike modules, libraries don't allow for dependencies, so when adding a
mix dep as a dependency, that dependency's mix.exs needs to be walked
and its deps as added deps to the original module recursively.
