+++ src/org/elixir_lang/mix/Dep.kt
@@ -34,8 +34,8 @@ data class Dep(val application: String, val path: String, val type: Type = Type.
                                 val key = keywordPair.keywordKey.text
 
                                 when (key) {
+                                    "app", "compile", "git", "github", "only", "optional" -> acc
                                     "in_umbrella" -> acc.copy(path =  "apps/$name", type = Type.MODULE)
-                                    "only" -> acc
                                     else -> {
                                         Logger.error(logger, "Don't know if Mix.Dep option `$key` is important for determining location of dependency", depsListElement)
                                         acc
